feat: implement 'ShouldQueue' interface to notifications
When a notification that implements 'ShouldQueue' is triggered, a job wraps up the notifications and is dispatched to the queue automatically. Also the name convention on notifications has changed: all of them have 'Notification' at the end of the name.
